refactoring: fix clang diagnostics III final

+++ Source/Android/V3DEngine/V3DIO/V3DBinaryRW.cpp
@@ -63,7 +63,7 @@ namespace V3D::V3DEngine::V3DIO
 		if (auto const asset = AAssetManager_open(static_cast<android_app*>(V3DEnvironment::GetApp())->activity->assetManager, fileFullName.ToChar(), AASSET_MODE_UNKNOWN))
 		{
 			const auto size = AAsset_getLength(asset);
-			char* buffer = V3DMemory::NewArray<char>(V3DFILE_INFO, (sizeof(char) * size) + 1);
+			char* buffer = V3DMemory::NewArray<char>(V3DFILE_INFO, (sizeof(char) * static_cast<unsigned int>(size)) + 1);
 
 			AAsset_read(asset, buffer, size);
 			result[size] = '\0';
@@ -78,7 +78,7 @@ namespace V3D::V3DEngine::V3DIO
 		return nullptr;
 	}
 	
-	void V3DBinaryRW::Write(const char* fileName, const char* data, long size, long offset)
+	void V3DBinaryRW::Write(const char* fileName, const char* data, unsigned int size, unsigned int offset)
 	{
 		V3DString fileFullName;
 		fileFullName += V3DString(static_cast<android_app*>(V3DEnvironment::GetApp())->activity->internalDataPath);
