+++ Source/Shared/V3DEngine/V3DIO/V3DMemoryStream.cpp
@@ -14,12 +14,12 @@ using namespace V3D::V3DEngine::V3DCore;
 
 namespace V3D::V3DEngine::V3DIO
 {
-	V3DMemoryStream::V3DMemoryStream(long maxSize) : isBufferAllocated{ true }, MaxSize{maxSize}
+	V3DMemoryStream::V3DMemoryStream(unsigned int maxSize) : isBufferAllocated{ true }, MaxSize{maxSize}
 	{
 		buffer = V3DMemory::NewArray<char>(V3DFILE_INFO, MaxSize);
 	}
 
-	V3DMemoryStream::V3DMemoryStream(long maxSize, char* bufferData) : isBufferAllocated{ false }, buffer{ bufferData }, MaxSize{ maxSize }
+	V3DMemoryStream::V3DMemoryStream(unsigned int maxSize, char* bufferData) : isBufferAllocated{ false }, buffer{ bufferData }, MaxSize{ maxSize }
 	{ }
 
 	V3DMemoryStream::V3DMemoryStream(char* bufferData) : isBufferAllocated{ false }, buffer{ bufferData }, MaxSize{ 0 }
@@ -38,18 +38,18 @@ namespace V3D::V3DEngine::V3DIO
 		return buffer;
 	}
 
-	long V3DMemoryStream::GetCurrentSize() const
+	unsigned int V3DMemoryStream::GetCurrentSize() const
 	{
 		return head;
 	}
 
-	void V3DMemoryStream::Write(const void* data, long dataSize)
+	void V3DMemoryStream::Write(const void* data, unsigned int dataSize)
 	{
 		memcpy(buffer + head, data, dataSize);
 		head += dataSize;
 	}
 
-	void V3DMemoryStream::Read(void* dataOut, long dataSize)
+	void V3DMemoryStream::Read(void* dataOut, unsigned int dataSize)
 	{
 		memcpy(dataOut, buffer + head, dataSize);
 		head += dataSize;
